Fix `BenchmarkParallelGetSet/redigo` test to actually use `redigo` pipelines. See [issue](https://github.com/mediocregopher/radix/issues/221).