@nussjustin 2020-07-20
The change is not correct, since the other 2 clients don't use manual pipelining either. The benchmark explicitly does not use pipelines for all 3 clients (redigo, redispipe, radix).

The benchmark is meant to show how implicit pipelining improves performance for concurrent goroutines. Both redispipe and radix will automatically pipeline concurrent commands under the hood as an optimization, which redigo does not do (at least it didn't the last time I checked). This allows redispipe and radix to be much faster when multiple goroutines try to run commands concurrently.

Optimizing the benchmark by using pipelines for the SET+GET may improve the performance per iteration, but doesn't make much sense considering the benchmarks goal. Also please note that (at least in my experience) people will often not use manual pipelining, even if possible and both redispipe and radix basically try to make manual pipelining for small pipelines completely unnecessary, so that people don't even have to think about it most of the time.

If we change the redigo benchmarks to use manual pipelining, then both the redispipe and radix benchmarks will have to be changed, too. Note that radix does not currently support manual pipelines with implicit pipelining (which is actually an explicit design decision to give more control over pipelining), which means that manual pipelining would probably result in a performance hit in this benchmark, but will probably in v4. redispipe also does not seem to support manual pipelining at all.