@bogdanciuca 2020-07-20
@nussjustin I think saying that "the change is not correct" is a bit of an overstatement. I understand the codewise advantage  of radix / redispipe - which automatically pipeline commands. 

> The benchmark is meant to show how implicit pipelining improves performance for concurrent goroutines. Both redispipe and radix will automatically pipeline concurrent commands under the hood as an optimization, which redigo does not do (at least it didn't the last time I checked). This allows redispipe and radix to be much faster when multiple goroutines try to run commands concurrently.

Why is implicit pipelining the benchmarks goal? Shouldn't it be general performance and a fair comparison? The benchmarks seems to show that `redigo` doesn't support pipelining and is outperformed (which is not true, if you write a few extra lines of code). I strongly feel this is the first wrong impression most people get form these benchmarks.