@nussjustin 2020-07-20
An extra benchmark should be ok. In fact, it'm quite interested in seeing how manual pipelining in radix compares to redigo :-)

The idea behind implicit pipelining is to reduce overhead (I/O and CPU) on both client and server without having to write manual pipelines (though even manual pipelines can benefit from implicit pipelining, just not in redispipe and radix since neither implements it). By combining commands from multiple concurrent goroutines we can batch both writes and reads, not only on the client, but also on the server, reducing the overhead of I/O (by doing less syscalls, which is by far the slowest part when using Redis) and even CPU usage.

While manual pipelining also reduces the overhead, it doesn't work as well for most cases, since both client and server still have to do the I/O and parsing per pipeline. We can easily see this when comparing the number of times the client and server have to do I/O and parsing, when comparing no pipelining with manual and implicit pipelining. For this I'm assuming 32 goroutines each doing a sequential SET + GET (similiar to the benchmark), but with only 1 iteration per goroutine:

* without pipelining (what the redigo code does) we pay 32*2 = 64 times for I/O and parsing on both client and server,
* with manual pipelining (what your modified redigo code does) we "only" pay 32 times for I/O and parsing and
* with implicit pipelining (what redispipe and radix do) in the base case we only pay 2 times for I/O and parsing, and that without any extra code!

And that is ignoring the fact, that redispipe and radix can do all that using only 1 connection, while redigo needs either multiple connections (which adds even more overhead) or needs to wait for each pipeline to finish before sending the next one.

Again we are testing what is at lest in my experience the common case of how people use Redis (many concurrent goroutines not using manual pipelining) and how redispipe and radix automatically optimize this.

The benchmark is meant to show how redispipe and radix reduce the overhead and improve the throughput and latency of your commands _for free_.

Plus since neither redispipe nor radix can use/combine manual pipelining with implicit pipelining, adding it just for redigo would make the comparison less fair (right know from a developers perspective all 3 benchmarks do the same thing, which makes them fair). Also given your stated improvement of 50%, redigo would still be much slower than redispipe and radix, thus not changing the point and only making the benchmark "unfair" in comparison.