@bogdanciuca 2020-07-20
> An extra benchmark should be ok. In fact, it'm quite interested in seeing how manual pipelining in radix compares to redigo :-)

Thanks, I think that should work. Any recommendations where to place it?

> - without pipelining (what the redigo code does) we pay 32*2 = 64 times for I/O and parsing on both client and server,
- with manual pipelining (what your modified redigo code does) we "only" pay 32 times for I/O and parsing and
- with implicit pipelining (what redispipe and radix do) in the base case we only pay 2 times for I/O and parsing, and that without any extra code!

So I am not trying to dismiss the benefits of implicit pipelining and its simplicity of use. I also think the batching optimisations available in `radix` and `redispipe` are amazing.

> Plus since neither redispipe nor radix can use/combine manual pipelining with implicit pipelining, adding it just for redigo would make the comparison less fair (right know from a developers perspective all 3 benchmarks do the same thing, which makes them fair). Also given your stated improvement of 50%, redigo would still be much slower than redispipe and radix, thus not changing the point and only making the benchmark "unfair" in comparison.

One could argue with the fact that the benchmarks do the same thing - since the `redigo` benchmark is not using any kind pipelining. It looks comparable with `BenchmarkParallelGetSet/radix/no_pipeline-8`. 

Even though it's still slower, `redigo` rises to `~80%` of `radix/redispipe` performance.

**Current benchmarks**:

`BenchmarkParallelGetSet/radix/no_pipeline-8                        38173             28154 ns/op              67 B/op          4 allocs/op
BenchmarkParallelGetSet/radix/one_pipeline-8                      106437             11153 ns/op              69 B/op          4 allocs/op
BenchmarkParallelGetSet/radix/default-8                           104139             11192 ns/op              70 B/op          4 allocs/op
BenchmarkParallelGetSet/redigo-8                                   **43832**             26623 ns/op             118 B/op          6 allocs/op
BenchmarkParallelGetSet/redispipe-8                               116005             10135 ns/op             168 B/op          8 allocs/op`

**PR benchmarks**:

`BenchmarkParallelGetSet/radix/no_pipeline-8                        41848             25495 ns/op              67 B/op          4 allocs/op
BenchmarkParallelGetSet/radix/one_pipeline-8                      104917             11082 ns/op              70 B/op          4 allocs/op
BenchmarkParallelGetSet/radix/default-8                           102685             11121 ns/op              70 B/op          4 allocs/op
BenchmarkParallelGetSet/redigo-8                                   **85759**             13636 ns/op             118 B/op          6 allocs/op
BenchmarkParallelGetSet/redispipe-8                               113139             10025 ns/op             168 B/op          8 allocs/op`